Add customized Iosevka font
Only regular and bold font weights, with upright and italic slopes, are
included in order to reduce the size of the build output, since those
are the only variations that actually see use.

To make it easier to build on non-NixOS machines: (I) The build plan is
written in TOML - which can be copied - instead of as a Nix expression;
and (II) The package is exposed as a flake output.